% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_lookup.R
\name{doi_lookup_to_duckdb}
\alias{doi_lookup_to_duckdb}
\title{Lookup DOIs in OpenAlex and save to a database in chunks of 1000}
\usage{
doi_lookup_to_duckdb(dois, dest = NULL)
}
\arguments{
\item{dois}{a character vector of DOIs}

\item{dest}{a path to an existing or wanted database file}
}
\value{
file path to the database file
}
\description{
Lookup DOIs in OpenAlex and save to a database in chunks of 1000
}
