% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openalex-package.R
\docType{package}
\name{openalex-package}
\alias{openalex}
\alias{openalex-package}
\title{openalex: Data from OpenAlex REST API}
\description{
The OpenAlex website provides open data on papers/works, venues, institutions and more around the world under the CC0 license. This R package provides some functions to access data from the OpenAlex REST API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kth-library.github.io/openalex/}
}

}
\author{
\strong{Maintainer}: Markus Skyttner \email{markussk@kth.se}

}
\keyword{internal}
