% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_alex_restclient.R
\name{openalex_attribution}
\alias{openalex_attribution}
\title{Attribution}
\usage{
openalex_attribution()
}
\description{
Use this attribution whenever data from the API is publicly displayed
}
\details{
OpenAlex provides a RESTful API for scholarly papers, authors,
institutions, and more. When publicly displaying data from the API,
it is polite to point back to OpenAlex at https://openalex.org/
}
