% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_alex_export.R
\name{openalex_counts}
\alias{openalex_counts}
\title{Counts from OpenAlex}
\usage{
openalex_counts(
  filter = openalex_filter_default(),
  dimensions = openalex_groupbys_default()
)
}
\arguments{
\item{filter}{a set of filter criteria, see the defaults in openalex_filter_default()}

\item{dimensions}{a set of grouping dimensions, see the defaults in openalex_groupbys_default()}
}
\value{
a list of tibbles
}
\description{
Aggregates/counts can be retrieved using the group_bys query parameter
}
