% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_lookup.R
\name{openalex_doi_lookup}
\alias{openalex_doi_lookup}
\title{Lookup DOIs using OpenAlex}
\usage{
openalex_doi_lookup(dois, resolution = c("all", "identifiers"))
}
\arguments{
\item{dois}{a character vector of DOIs}

\item{resolution}{either "all" or "identifiers" to only return other related identifiers}
}
\value{
tibble(s)
}
\description{
Lookup DOIs using OpenAlex
}
