% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_alex_restclient.R
\name{openalex_flatten_long}
\alias{openalex_flatten_long}
\title{Flatten R object from deserialized nested JSON object}
\usage{
openalex_flatten_long(nestedlist)
}
\arguments{
\item{nestedlist}{a nested list of lists}
}
\value{
a tibble in long format
}
\description{
Flatten R object from deserialized nested JSON object
}
