% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_alex_restclient.R
\name{openalex_key}
\alias{openalex_key}
\title{Use an API key for OpenAlex Premium Subscription}
\usage{
openalex_key(key)
}
\arguments{
\item{key}{a premium subscription key}
}
\value{
a logical depending on whether key was set or unset
}
\description{
This provides access to the latest data, fresher than what snapshots provide.
It also enables faster requests and filtering on from_created_date and from_updated_date fields.
}
\details{
Additional details...

\url{https://github.com/ourresearch/openalex-api-tutorials/blob/main/notebooks/getting-started/premium.ipynb}
\url{https://docs.openalex.org/api-entities/works/filter-works#from_created_date}
\url{https://docs.openalex.org/api-entities/works/filter-works#from_updated_date}
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
 openalex_key("my_secret_api_key")
 openalex_key("")
\dontshow{\}) # examplesIf}
}
