% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_alex_restclient.R
\name{openalex_kth_rawaff_query}
\alias{openalex_kth_rawaff_query}
\title{Example query when searching raw affiliation strings}
\usage{
openalex_kth_rawaff_query()
}
\value{
string with query
}
\description{
This variant is specifically tailored for KTH, Royal Institute of Technology
and includes some affiliation string variations which might be related.
}
