% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_lookup.R
\name{openalex_orcid_lookup}
\alias{openalex_orcid_lookup}
\title{Lookup ORCiDs using OpenAlex}
\usage{
openalex_orcid_lookup(orcids)
}
\arguments{
\item{orcids}{a character vector of DOIs}
}
\value{
tibble(s)
}
\description{
Lookup ORCiDs using OpenAlex
}
