% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_alex_restclient.R
\name{openalex_polite}
\alias{openalex_polite}
\title{Enter the OpenAlex API polite pool for faster requests by providing an email}
\usage{
openalex_polite(email)
}
\arguments{
\item{email}{an email address, on the form "you@example.com" or "" to unset email}
}
\value{
a logical depending on whether email was set or unset
}
\description{
Enter the OpenAlex API polite pool for faster requests by providing an email
}
\examples{
\dontrun{
if(interactive()){
 # to set
 openalex_polite("you@example.com")
 # to unset
 openalex_polite("")
 }
}
}
