% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_alex_restclient.R
\name{openalex_work}
\alias{openalex_work}
\title{Retrieve work from OpenAlex REST API}
\usage{
openalex_work(identifier, format = "table", use_random = FALSE)
}
\arguments{
\item{identifier}{string with identifier}

\item{format}{one of "table" or "object"}

\item{use_random}{logical to indicate whether to use random identifier, Default: FALSE}
}
\value{
as per format, either a tibble or an R object
}
\description{
This function retrieves works given an identifier
}
\examples{
\dontrun{
 openalex_work(use_random = TRUE)
 }
}
