% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cursor_crawl.R
\name{openalex_works_cursorcrawl}
\alias{openalex_works_cursorcrawl}
\title{Crawl multipage responses from queries against the API}
\usage{
openalex_works_cursorcrawl(works_filter, n_max_pages = 5)
}
\arguments{
\item{works_filter}{the works filter}

\item{n_max_pages}{the max amount of pages to fetch (50 per page)}
}
\value{
paths to downloaded files
}
\description{
Chunks and uses cursor based pagination to fetch works
}
