% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_alex_export.R
\name{openalex_works_export}
\alias{openalex_works_export}
\title{Use OpenAlex API for exporting data in tabular and wos formats}
\usage{
openalex_works_export(q, fmt = c("csv", "wos-plaintext"), raw_string = FALSE)
}
\arguments{
\item{q}{the query, for example "authorships.institutions.lineage:i86987016,authorships.institutions.lineage:!i4210161097,type:types/article,primary_location.source.type:source-types/journal|source-types/conference,publication_year:2023"}

\item{fmt}{the export format, one of "csv" or "wos-plaintext" or "wos-plaintext-diva"}

\item{raw_string}{boolean to indicate whether a raw string should be returned}
}
\value{
a character vector with a raw string with the results from the export or a data frame
}
\description{
Use OpenAlex API for exporting data in tabular and wos formats
}
