% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_alex_restclient.R
\name{openalex_works_published_since}
\alias{openalex_works_published_since}
\title{Recently published works based on query for matching raw affiliations}
\usage{
openalex_works_published_since(
  raw_search_criteria = openalex_kth_rawaff_query(),
  since_days = 7
)
}
\arguments{
\item{raw_search_criteria}{raw affiliation string search criteria,
by default openalex_kth_rawaff_query()}

\item{since_days}{integer indicating days back from today}
}
\value{
list of tables with results
}
\description{
Recently published works based on query for matching raw affiliations
}
