% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_alex_restclient.R
\name{openalex_works_updated_since}
\alias{openalex_works_updated_since}
\title{Recently updated works based on query for matching raw affiliations}
\usage{
openalex_works_updated_since(
  raw_search_criteria = openalex_kth_rawaff_query(),
  since_minutes
)
}
\arguments{
\item{raw_search_criteria}{raw affiliation string search criteria,
by default openalex_kth_rawaff_query()}

\item{since_minutes}{integer indicating minutes since now}
}
\value{
list of tables with results
}
\description{
This function requires a premium subscription API key to be set.
}
