% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_alex_export.R
\name{openalex_write_duckdb}
\alias{openalex_write_duckdb}
\title{Export the results from a crawl as a duckdb database file}
\usage{
openalex_write_duckdb(crawl, destdir = NULL)
}
\arguments{
\item{crawl}{the results from running the to_tbls fcn}

\item{destdir}{the location to save the database file}
}
\value{
file path to the database file
}
\description{
Export the results from a crawl as a duckdb database file
}
