% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid_lookup.R
\name{orcid_lookup_to_duckdb}
\alias{orcid_lookup_to_duckdb}
\title{Lookup ORCiDs in OpenAlex and save to a database in chunks of 100}
\usage{
orcid_lookup_to_duckdb(orcids, dest = NULL)
}
\arguments{
\item{orcids}{a character vector of DOIs}

\item{dest}{a path to an existing or wanted database file}
}
\value{
file path to the database file
}
\description{
Lookup ORCiDs in OpenAlex and save to a database in chunks of 100
}
