% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openalex-package.R
\docType{data}
\name{topics}
\alias{topics}
\title{Topics}
\format{
A data frame with 4516 rows and 9 variables:
\describe{
\item{\code{id_topic}}{character the id for the topic}
\item{\code{topic}}{character description of topic}
\item{\code{description}}{character long form description of this topic cluster}
\item{\code{id_subfield}}{character the id for the subfield of this topic}
\item{\code{subfield}}{character description of the subfield}
\item{\code{id_field}}{character the id of the field}
\item{\code{field}}{character description of the field}
\item{\code{id_domain}}{character the id of the domain}
\item{\code{domain}}{character description of the domain}
}
}
\usage{
topics
}
\description{
Topics used by OpenAlex
}
\details{
DETAILS
}
\keyword{datasets}
