% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_alex_export.R
\name{wos_plaintext_for_diva}
\alias{wos_plaintext_for_diva}
\title{Function which converts a wos_plaintext-string into a format
which can be uploaded to DiVA, by adding ER tags
(including a blank line) after each record}
\usage{
wos_plaintext_for_diva(x)
}
\arguments{
\item{x}{character string with "wos-plaintext" format as returned from OpenAlex export API endpoint}
}
\description{
Function which converts a wos_plaintext-string into a format
which can be uploaded to DiVA, by adding ER tags
(including a blank line) after each record
}
