% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_alex_restclient.R
\name{openalex_crawl}
\alias{openalex_crawl}
\title{Crawl multiple pages of results}
\usage{
openalex_crawl(entity, query, verbose = FALSE, fmt = "object")
}
\arguments{
\item{entity}{one of the values in openalex_entity_enum()}

\item{query}{an openalex_query object}

\item{verbose}{boolean to indicate whether to output messages during process}

\item{fmt}{the return format, one of "object" or "tables"}
}
\value{
R object with results matching the query
}
\description{
Iterates over paged results showing a progress bar
}
